Fix loss of state of navigation
This commit fixes a really important bug that I introduced in my implementation of navigation.

When we navigated to a new screen, "Add workout" all of the state of the old ViewModel was lost. This is really, really bad, because the whole point of that screen is that you will return to the "Create intervals" screen (currently TimerScreen) to add more to what currently exists.